+++ packages/grid/src/index.ts
@@ -1,4 +1,5 @@
 export { type BinaryGrid } from './binary-grid';
+export { getRegion, getRegions, type Cell } from './flood-fill';
 export { getNeighbours } from './get-neighbours';
 export { randomGrid, type FillOptions, type GridOptions } from './random-grid';
 export { randomWalker, type WalkerOptions } from './random-walker';